new(living-frontend) Re-written predefined 'structures'
Re-wrote the pre-defined structures mechanism to use the native livingdocs
serialization format to enable Duplication of components.

Note : For _old_ designs that have structures, rename the "type"
key with "identifier" and change the "type" value to include a prefix
of the design's name, ie "bootstrap3.oldtype". Next, change the inner
"components" key to "containers" - the top level one (ie label, components)
can remain as-is.